# tests/hosts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AssistantFeedbackObservation": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "content_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "event_type": {
            "const": "assistant_feedback",
            "default": "assistant_feedback",
            "title": "Event Type",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 384,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "observation_id": {
            "pattern": "^obs_[0-9a-f]{32}$",
            "title": "Observation Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/AssistantFeedbackPayload"
          },
          "provenance": {
            "$ref": "#/components/schemas/ObservationProvenance"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "observation_id",
          "idempotency_key",
          "occurred_at",
          "received_at",
          "provenance",
          "payload"
        ],
        "title": "AssistantFeedbackObservation",
        "type": "object"
      },
...
      "ContentKind": {
        "additionalProperties": false,
        "description": "Provider-declared stable content-kind identity.",
        "properties": {
          "value": {
            "pattern": "^[a-z][a-z0-9_-]{0,63}$",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "ContentKind",
        "type": "object"
      },
      "ContentOpenedObservation": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "content_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "event_type": {
            "const": "content_opened",
            "default": "content_opened",
            "title": "Event Type",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 384,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "observation_id": {
            "pattern": "^obs_[0-9a-f]{32}$",
            "title": "Observation Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/HostOpenPayload"
          },
          "provenance": {
            "$ref": "#/components/schemas/ObservationProvenance"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "observation_id",
          "idempotency_key",
          "occurred_at",
          "received_at",
          "provenance",
          "payload"
        ],
        "title": "ContentOpenedObservation",
        "type": "object"
      },
...
      "ContentRef": {
        "additionalProperties": false,
        "description": "Canonical cross-process identity for one provider-native record.\n\nURL normalization remains provider-owned. This layer validates and stores\nthe provider's canonical HTTP(S) result without rewriting it.",
        "properties": {
          "canonical_url": {
            "maxLength": 2048,
            "pattern": "^https?://[^\\s]+$",
            "title": "Canonical Url",
            "type": "string"
          },
          "content_kind": {
            "$ref": "#/components/schemas/ContentKind"
          },
          "provider_content_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Provider Content Id",
            "type": "string"
          },
          "provider_id": {
            "$ref": "#/components/schemas/ProviderId"
          }
        },
        "required": [
          "provider_id",
          "content_kind",
          "provider_content_id",
          "canonical_url"
        ],
        "title": "ContentRef",
        "type": "object"
      },
...
      "ContentSavedObservation": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "content_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "event_type": {
            "const": "content_saved",
            "default": "content_saved",
            "title": "Event Type",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 384,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "observation_id": {
            "pattern": "^obs_[0-9a-f]{32}$",
            "title": "Observation Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/EmptyPayload"
          },
          "provenance": {
            "$ref": "#/components/schemas/ObservationProvenance"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "observation_id",
          "idempotency_key",
          "occurred_at",
          "received_at",
          "provenance",
          "payload"
        ],
        "title": "ContentSavedObservation",
        "type": "object"
      },
...
      "DeterministicProfileEditObservation": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "content_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "event_type": {
            "const": "deterministic_profile_edit",
            "default": "deterministic_profile_edit",
            "title": "Event Type",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 384,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "observation_id": {
            "pattern": "^obs_[0-9a-f]{32}$",
            "title": "Observation Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ProfileEditPayload"
          },
          "provenance": {
            "$ref": "#/components/schemas/ObservationProvenance"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "observation_id",
          "idempotency_key",
          "occurred_at",
          "received_at",
          "provenance",
          "payload"
        ],
        "title": "DeterministicProfileEditObservation",
        "type": "object"
      },
...
      "EmptyPayload": {
        "additionalProperties": false,
        "properties": {},
        "title": "EmptyPayload",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "maxLength": 500,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "FeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "content_ref": {
            "$ref": "#/components/schemas/ContentRef"
          },
          "dwell_ms": {
            "anyOf": [
              {
                "maximum": 86400000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dwell Ms"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "opened",
              "liked",
              "disliked",
              "saved",
              "dismissed"
            ],
            "title": "Kind",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "shown_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Shown Id",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key",
          "shown_id",
          "content_ref",
          "kind"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
      "FeedbackResponse": {
        "additionalProperties": false,
        "properties": {
          "result": {
            "$ref": "#/components/schemas/RecordFeedbackResult"
          }
        },
        "required": [
          "result"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
...
      "ObservationsRequest": {
        "additionalProperties": false,
        "properties": {
          "allowed_event_types": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Allowed Event Types",
            "type": "array"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "observations": {
            "items": {
              "discriminator": {
                "mapping": {
                  "assistant_feedback": "#/components/schemas/AssistantFeedbackObservation",
                  "content_opened": "#/components/schemas/ContentOpenedObservation",
                  "content_saved": "#/components/schemas/ContentSavedObservation",
                  "deterministic_profile_edit": "#/components/schemas/DeterministicProfileEditObservation",
                  "preference_statement": "#/components/schemas/PreferenceStatementObservation",
                  "provider_history_import": "#/components/schemas/ProviderHistoryImportObservation",
                  "recommendation_disliked": "#/components/schemas/RecommendationDislikedObservation",
                  "recommendation_dismissed": "#/components/schemas/RecommendationDismissedObservation",
                  "recommendation_liked": "#/components/schemas/RecommendationLikedObservation",
                  "recommendation_opened": "#/components/schemas/RecommendationOpenedObservation",
                  "recommendation_saved": "#/components/schemas/RecommendationSavedObservation",
                  "recommendation_shown": "#/components/schemas/RecommendationShownObservation"
                },
                "propertyName": "event_type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RecommendationShownObservation"
                },
                {
                  "$ref": "#/components/schemas/RecommendationOpenedObservation"
                },
                {
                  "$ref": "#/components/schemas/RecommendationLikedObservation"
                },
                {
                  "$ref": "#/components/schemas/RecommendationDislikedObservation"
                },
                {
                  "$ref": "#/components/schemas/RecommendationSavedObservation"
                },
                {
                  "$ref": "#/components/schemas/RecommendationDismissedObservation"
                },
                {
                  "$ref": "#/components/schemas/ContentOpenedObservation"
                },
                {
                  "$ref": "#/components/schemas/ContentSavedObservation"
                },
                {
                  "$ref": "#/components/schemas/AssistantFeedbackObservation"
                },
                {
                  "$ref": "#/components/schemas/PreferenceStatementObservation"
                },
                {
                  "$ref": "#/components/schemas/DeterministicProfileEditObservation"
                },
                {
                  "$ref": "#/components/schemas/ProviderHistoryImportObservation"
                }
              ]
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Observations",
            "type": "array"
          }
        },
        "required": [
          "idempotency_key",
          "observations",
          "allowed_event_types"
        ],
        "title": "ObservationsRequest",
        "type": "object"
      },
      "ObservationsResponse": {
        "additionalProperties": false,
        "properties": {
          "result": {
            "$ref": "#/components/schemas/RecordBatchResult"
          }
        },
        "required": [
          "result"
        ],
        "title": "ObservationsResponse",
        "type": "object"
      },
...
      "ProviderId": {
        "additionalProperties": false,
        "description": "Stable lower-case provider identity; never a display name.",
        "properties": {
          "value": {
            "pattern": "^[a-z][a-z0-9_-]{0,63}$",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "ProviderId",
        "type": "object"
      },
...
      "RecommendationPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SelectionRecord"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RecommendationPage",
        "type": "object"
      },
...
      "RecordBatchResult": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecordItemResult"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RecordBatchResult",
        "type": "object"
      },
      "RecordFeedbackResult": {
        "additionalProperties": false,
        "properties": {
          "feedback_id": {
            "title": "Feedback Id",
            "type": "string"
          },
          "inserted": {
            "title": "Inserted",
            "type": "boolean"
          },
          "observation_id": {
            "title": "Observation Id",
            "type": "string"
          }
        },
        "required": [
          "feedback_id",
          "observation_id",
          "inserted"
        ],
        "title": "RecordFeedbackResult",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "maximum_items": {
            "default": 50,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Maximum Items",
            "type": "integer"
          }
        },
        "required": [
          "idempotency_key"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RefreshResponse": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "title": "Decision",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "RefreshResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenBiliClaw API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/feedback": {
      "post": {
        "operationId": "feedback_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
    "/v1/observations": {
      "post": {
        "operationId": "observations_v1_observations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Observations",
        "tags": [
          "feedback"
        ]
      }
    },
...
    "/v1/recommendations": {
      "get": {
        "operationId": "recommendations_v1_recommendations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/v1/recommendations/refresh": {
      "post": {
        "operationId": "refresh_v1_recommendations_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Refresh",
        "tags": [
          "recommendations"
        ]
      }
    },
...
  }
}